Support terminals lacking different cursor visibilities
Some terminals (notably older aterm/kterm) don't support changing the
cursor visibility. This patch simply ignores failures to do so. This is
not a problem in any of the available screen modes: In the stats views
the visible cursor will remain bottom right after drawing the screen and
in the CLI view the cursor is active on the input line anyways.

Signed-off-by: John Feuerstein <john@feurix.com>