Prevent infinite size sync if screen is larger than supported
There was a corner case where a dumbtty and a huge screen size could
cause infinite size syncs (and thus, massive screen flicker).

Note that with Python 2.4 all terminals are considered dumb because of
missing SIGWINCH and resize_term support in the curses module.

Thanks to Jérémy Bonnet for reporting this.